Use constant-sized integers for storing parameters
We get data from v8 as uint32_t and pass it to argon2 as uint32_t, so
we might as well store it as uint32_t. While unsigned int is almost
universally 32 bits these days, it's only guaranteed to be at least
16 bits.

This is probably isn't an important consideration for time_cost and
parallelism, but the memory cost is stored as 2^x, with x=12 by
default. At x=16, memory_cost would otherwise overflow on LP32
platforms.