Improve bad salt handling
Node's Buffer::toString interprets the data as UTF-8 by default. This
results in decoding errors when trying to convert the random bytes to
string, because some byte sequences are not valid unicode. The
replacement of invalid unicode with the Unicode Replacement Character
(U+FFFD) greatly decreases the entropy of the string and significantly
lengthens its representation in bytes. The later truncation of the
string to 16 bytes then further decreases its entropy.

Specifying that the string is ascii reduces entropy somewhat, as null
bytes are converted to spaces, and the high bit is stripped from each
byte. This is still better than the utf-8 errors, but not perfect.

Specifying the string as binary, unfortunately, results in a longer
representation when finally converted back to UTF-8 for hashing. There
may be a way to handle this in the C++ code, but it's unclear how you'd
identify the encoding given that we can't depend on the salt coming
from our own code.

Within the hash function itself, the use of std::strlen for determining
the length of the salt is inappropriate, as completely random strings
may contain null bytes.

There is a real question as to whether the password or the salt should
actually even be strings, but that can't be changed without breaking
backwards compatibility. Ideally, we'd just keep everything in byte
buffers and not worry about string encodings at all.